% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WhatsappEmoji.R
\name{emoji_loader}
\alias{emoji_loader}
\title{Load Emojis}
\usage{
emoji_loader(filename)
}
\arguments{
\item{filename}{, string}
}
\description{
This function loads the emojis from csv file.
}
\examples{
emoji_all = emoji_loader("whatsapp_emoji_all.csv")
}

