% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WhatsappEmoji.R
\name{emoji_replacer}
\alias{emoji_replacer}
\title{Replace Emoji with name}
\usage{
emoji_replacer(data)
}
\arguments{
\item{data}{, data frame}
}
\description{
This function replaces all the emojis with their respective names.
}
\examples{
clean_data_without_media <- emoji_replacer(emoji_all)
}

