% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WhatsappEmoji.R
\name{remove_unused_emoji}
\alias{remove_unused_emoji}
\title{Remove unused emojis}
\usage{
remove_unused_emoji(data_emoji)
}
\arguments{
\item{data_emoji}{, data frame}
}
\description{
This function removes all emoji columns that have sum equal to zero.
}
\examples{
clean_data_without_media_with_emoji <- remove_unused_emoji(clean_data_without_media_with_emoji)
}

