% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WhatsappEmoji.R
\name{emoji_human_color_ignore}
\alias{emoji_human_color_ignore}
\title{Human Emojis}
\usage{
emoji_human_color_ignore(data)
}
\arguments{
\item{data}{, data frame}
}
\description{
This function is used to ignore the skin tones and considers all colors as same.
}
\examples{
emoji_human <- emoji_human_color_ignore(emoji_human)
}

